+++ llvm/cmake/modules/CheckCompilerVersion.cmake
@@ -10,7 +10,10 @@ set(CLANG_MIN 3.1)
 set(CLANG_SOFT_ERROR 3.5)
 set(APPLECLANG_MIN 3.1)
 set(APPLECLANG_SOFT_ERROR 6.0)
-set(MSVC_MIN 19.00.24213.1)
+
+# https://en.wikipedia.org/wiki/Microsoft_Visual_C#Internal_version_numbering
+# _MSC_VER == 1910 MSVC++ 14.1 (Visual Studio 2017 version 15.0)
+set(MSVC_MIN 19.1)
 set(MSVC_SOFT_ERROR 19.1)
 
 # Map the above GCC versions to dates: https://gcc.gnu.org/develop.html#timeline